refactor migration, automated createdupdatedby using traits, remove search not found and update pagination

+++ database/migrations/2023_07_15_121233_create_vendor_limits_table.php
@@ -13,22 +13,12 @@ return new class extends Migration
     {
         Schema::create('vendor_limits', function (Blueprint $table) {
             $table->id();
-            // vendor_id
-            // $table->foreignId('vendor_id')->nullable();
-            // limit
             $table->string('name');
             $table->integer('amount_limit');
-
-            // relation
-            // $table->foreign('vendor_id')->references('id')->on('vendors')->onDelete('cascade');
-            // created by
-            $table->integer('created_by')->nullable();
-            // updated by
-            $table->integer('updated_by')->nullable();
-            // deleted by
-            $table->integer('deleted_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('deleted_by')->nullable();
             $table->timestamps();
-            // soft delete
             $table->softDeletes();
 
             // index
