+++ database/migrations/2023_07_15_133620_create_memberships_table.php
@@ -13,24 +13,15 @@ return new class extends Migration
     {
         Schema::create('memberships', function (Blueprint $table) {
             $table->id();
-            // icon
             $table->string('image')->nullable();
-            // name
             $table->string('name');
-            // from
             $table->integer('from');
-            // to
             $table->integer('until');
-            // point
             $table->integer('point');
-            // created by
-            $table->integer('created_by')->nullable();
-            // updated by
-            $table->integer('updated_by')->nullable();
-            // deleted by
-            $table->integer('deleted_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('deleted_by')->nullable();
             $table->timestamps();
-            // soft delete
             $table->softDeletes();
 
             // index
