+++ database/migrations/2023_07_29_231036_create_product_variants_table.php
@@ -13,17 +13,12 @@ return new class extends Migration
     {
         Schema::create('product_variants', function (Blueprint $table) {
             $table->id();
-            // relation to product attribute
             $table->foreignId('product_attribute_id')->constrained('product_attributes')->onDelete('cascade');
             $table->string('name');
-            // created by
-            $table->integer('created_by')->nullable();
-            // updated by
-            $table->integer('updated_by')->nullable();
-            // deleted by
-            $table->integer('deleted_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('deleted_by')->nullable();
             $table->timestamps();
-            // soft delete
             $table->softDeletes();
         });
     }
