+++ database/migrations/2023_07_29_232044_create_checklist_items_table.php
@@ -13,17 +13,12 @@ return new class extends Migration
     {
         Schema::create('checklist_items', function (Blueprint $table) {
             $table->id();
-            // relation to checklist category
             $table->foreignId('checklist_category_id')->constrained('checklist_categories')->onDelete('cascade');
             $table->string('name');
-            // created by
-            $table->integer('created_by')->nullable();
-            // updated by
-            $table->integer('updated_by')->nullable();
-            // deleted by
-            $table->integer('deleted_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('deleted_by')->nullable();
             $table->timestamps();
-            // soft delete
             $table->softDeletes();
         });
     }
