+++ database/migrations/2023_08_06_061855_create_vehicles_table.php
@@ -15,14 +15,10 @@ return new class extends Migration
             $table->id();
             $table->string('model_name');
             $table->string('plate_number');
-            // created by
-            $table->integer('created_by')->nullable();
-            // updated by
-            $table->integer('updated_by')->nullable();
-            // deleted by
-            $table->integer('deleted_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('deleted_by')->nullable();
             $table->timestamps();
-            // soft delete
             $table->softDeletes();
         });
     }
